Add files via upload
VerifyCityPath application determines if two cities are connected. Two cities are considered
connected if there’s a series of roads that can be traveled from one city to another.

A rest api is exposed which determines the possible path between two cities. If path exists then return "Yes" or "No" in response